requirements.txt
flask
streamlit
cryptography
pillow
